Removed BaseObserve pattern
